Able to save critics settings on backend
