+++ Arkumida/webapi/Models/Api/Responses/Creatures/ChangeEmailResponse.cs
@@ -18,7 +18,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace webapi.Models.Api.Responses;
+namespace webapi.Models.Api.Responses.Creatures;
 
 /// <summary>
 /// Email change response
