+++ Arkumida/webapi/Models/Api/Responses/Creatures/ChangePasswordResponse.cs
@@ -18,7 +18,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace webapi.Models.Api.Responses;
+namespace webapi.Models.Api.Responses.Creatures;
 
 /// <summary>
 /// Response with information about changed password
