+++ Arkumida/webapi/Models/Api/Responses/Creatures/InitiateEmailChangeResponse.cs
@@ -18,7 +18,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace webapi.Models.Api.Responses;
+namespace webapi.Models.Api.Responses.Creatures;
 
 /// <summary>
 /// Result of email change initiation
