+++ Arkumida/webapi/Models/Api/Responses/Creatures/PasswordResetResponse.cs
@@ -18,7 +18,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace webapi.Models.Api.Responses;
+namespace webapi.Models.Api.Responses.Creatures;
 
 /// <summary>
 /// Password reset confirmation response
